requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cached-property==1.5.2
cliar==1.3.5
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mypy-extensions==1.0.0
pillow==11.0.0
pip-versions==0.2.0
SQLAlchemy==2.0.36
sqlalchemy-orm==1.2.10
typeguard==4.4.2
typing-inspect==0.9.0
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
